# dashboard/API_CONTRACT.yaml
openapi: 3.0.3
info:
  title: MT5 Trading Dashboard API
  description: |
    API REST para gerenciamento de bot de trading MT5.
    Backend Flask deve implementar estes endpoints para integração com o frontend React.
  version: 1.0.0
  contact:
    name: MT5 Trading Dashboard
    email: support@example.com
servers:
  - url: http://localhost:5000
    description: Servidor de desenvolvimento local
  - url: https://api.trading-bot.com
    description: Servidor de produção
security:
  - BearerAuth: []
...
paths:
  /api/login:
    post:
      tags:
        - Authentication
      summary: Login de administrador
      description: Autentica usuário e retorna token JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  format: password
                  example: admin123
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  expires_in:
                    type: integer
                    example: 3600
                  user:
                    type: object
                    properties:
                      username:
                        type: string
                      role:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/account:
    get:
      tags:
        - Account
      summary: Informações da conta
      description: Retorna dados da conta MT5 (saldo, patrimônio, margem)
      responses:
        '200':
          description: Dados da conta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/strategies/toggle:
    post:
      tags:
        - Strategies
      summary: Ativar/desativar estratégia
      description: Liga ou desliga uma estratégia específica
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - strategy
                - enabled
              properties:
                strategy:
                  type: string
                  example: RSI_Strategy
                enabled:
                  type: boolean
      responses:
        '200':
          description: Estratégia atualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  strategy:
                    type: string
                  enabled:
                    type: boolean
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Account:
      type: object
      properties:
        balance:
          type: number
          format: double
          example: 10000.00
        equity:
          type: number
          format: double
          example: 10250.50
        margin:
          type: number
          format: double
          example: 500.00
        free_margin:
          type: number
          format: double
          example: 9750.50
        margin_level:
          type: number
          format: double
          example: 2050.10
        profit:
          type: number
          format: double
          example: 250.50
        currency:
          type: string
          example: USD
        leverage:
          type: integer
          example: 100
        server:
          type: string
          example: MetaQuotes-Demo
        name:
          type: string
          example: John Doe
        number:
          type: integer
          example: 12345678
...
  responses:
    Unauthorized:
      description: Token inválido ou ausente
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
...